Functional commit. This commit is solely a functional working, not a refactored one.
