+++ src/Components/CharacterList.js
@@ -4,8 +4,6 @@ import Character from './Character'
 
 const CharacterList = (props) => {
 
-
- 
   return (
     <div> 
        {props.list.map((data) => <Character   name={data.name} birthYear={data.birth_year} height={data.height} mass={data.mass} homeWorld={data.homeWorld} species={data.species} key={data.name} />
